Added ReadReq() which ensures the CLOCK register synchronised.
Reading the CLOCK elements when it is not synchronised can cause
the bus to stall. This can block interrupts. This code currently increases
the cost of calling getEpoch() from 6ms to about 35ms.

Todo: Change getEpoch() to single read from the CLOCK register. This
will reduce the cost back to ~6ms.