The library now attempts to preserve the clock time value.
It tests whether the reset cause was not due to POR or BOD (either).
Additionally, it checks whether the clock was already configured
in MODE2 (calendar mode). If both of these were the case it saves and
the existing clock value and restores it after the configuration process
has been completed.

An optional parameter to begin() will always reset the clock time.
This is default false, and it will by default attempt to preserve the clock time.