Added support for Mode 0 and Mode 1
Adds support for Mode 0 (32-bit counter) and Mode 1 (16-bit periodic counter).  Mode 1 can operate with one or two compare values and optionally a set period.  Adds options to reset on match and interrupt on overflow as well as a helper function to get the source of RTC interrupts and enums for prescale values and interrupt sources.  Includes examples for Mode 0 with reset on match, Mode 1 with two compares and a set period and Mode 1 with interrupt on overflow enabled (along with two compares).